move status and auth to domains
