+++ server/src/controllers/index.ts
@@ -1,13 +1,9 @@
-import AuthControllerHTTP from "./authController.js";
 import ChannelControllerHTTP from "./channelController.js";
 import MessageControllerHTTP from "./messageController.js";
 import RelationshipControllerHTTP from "./relationshipController.js";
-import StatusControllerHTTP from "./statusController.js";
 
 export {
-  AuthControllerHTTP,
   ChannelControllerHTTP,
   MessageControllerHTTP,
   RelationshipControllerHTTP,
-  StatusControllerHTTP,
 };
