+++ server/src/domains/auth/index.ts
@@ -1,0 +1,2 @@
+export { AuthControllerHTTP } from "./controllers/http/auth.js";
+export { AuthService } from "./services/auth.js";
