+++ server/src/routes/usersRouter.ts
@@ -2,7 +2,7 @@ import { Router } from "express";
 import validate from "../middlewares/validate.js";
 import authGuard from "../middlewares/authGuard.js";
 import { getUserProfileByIdSchema } from "../domains/users/validators/user.js";
-import { patchUserStatus } from "../validators/statusValidators.js";
+import { patchUserStatus } from "../domains/users/validators/status.js";
 import { userControllerHTTP } from "../instances.js";
 import { statusControllerHTTP } from "../instances.js";
