+++ server/src/services/index.ts
@@ -1,13 +1,5 @@
-import AuthService from "./authService.js";
 import MessageService from "./messageService.js";
-import StatusService from "./statusService.js";
 import RelationshipService from "./relationshipService.js";
 import NotificationService from "./notificationService.js";
 
-export {
-  AuthService,
-  MessageService,
-  StatusService,
-  RelationshipService,
-  NotificationService,
-};
+export { MessageService, RelationshipService, NotificationService };
